scraper/ig_feed_scraper/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ig-feed-scraper"
version = "0.1.0"
description = "Instagram feed scraper using Instaloader"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instaloader>=4.10.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
py-modules = ["api", "ig_scraper"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
